Create classes for package generation
