+++ turtle_nest/include/turtle_nest/package_generators/python_package_generator.h
@@ -1,5 +1,5 @@
 /* ------------------------------------------------------------------
- * Copyright 2024 Janne Karttunen
+ * Copyright 2025 Janne Karttunen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,29 +14,25 @@
  * limitations under the License.
  * ------------------------------------------------------------------
 */
+#ifndef PYTHON_PACKAGE_GENERATOR_H
+#define PYTHON_PACKAGE_GENERATOR_H
 
-#ifndef GENERATE_NODE_H
-#define GENERATE_NODE_H
+#include "turtle_nest/package_generators/base_package_generator.h"
 
-#include <QString>
-#include "turtle_nest/packageinfo.h"
 
+class PythonPackageGenerator : public BasePackageGenerator
+{
+public:
+  void add_node(QString node_name, NodeType node_type, QString package_path, QString package_name) override;
+};
 
 void generate_python_node(
-  QString package_path, QString package_name, QString node_name,
-  bool create_config, bool overwrite_existing = false);
+    QString package_path, QString package_name, QString node_name,
+    bool create_config, bool overwrite_existing = false);
 void create_init_file(QString package_path, QString package_name);
 void add_exec_permissions(QString node_path);
-void generate_cpp_node(
-  QString package_path, QString node_name,
-  bool create_config, bool overwrite_existing = false);
-void add_node_to_cmakelists(PackageInfo pkg_info, QString node_name);
-void add_dependency_to_cmakelists(QString dependency, QString cmakelists_path);
 void add_rclpy_dependency_to_package_xml(QString package_path);
-void add_rclcpp_dependency_to_package_xml(QString package_path);
-void install_python_modules_in_cmakelists(QString cmakelists_path);
-void add_python_node_to_cmakelists(QString package_path, QString node_name);
-void add_node_to_setup_py(PackageInfo pkg_info, QString node_name);
-QString generate_new_setup_py(PackageInfo pkg_info, QString node_name);
+void add_node_to_setup_py(QString package_path, QString package_name, QString node_name);
+QString generate_new_setup_py(QString package_path, QString package_name,  QString node_name);
 
-#endif // GENERATE_NODE_H
+#endif // PYTHON_PACKAGE_GENERATOR_H
